refactor(persistence): move usage bucketers + schedule-attribution SQL into persistence/
