+++ assistant/src/runtime/routes/usage-routes.ts
@@ -9,7 +9,6 @@
 
 import { z } from "zod";
 
-import { validateTimezone } from "../../memory/usage-buckets.js";
 import {
   getUsageDayBuckets,
   getUsageGroupBreakdown,
@@ -22,6 +21,7 @@ import {
   type UsageAggregationFilter,
   type UsageGranularity,
 } from "../../persistence/llm-usage-store.js";
+import { validateTimezone } from "../../persistence/usage-buckets.js";
 import { ACTOR_PRINCIPALS } from "../auth/route-policy.js";
 import { parseEpochMillisRange } from "./epoch-millis-range.js";
 import { BadRequestError } from "./errors.js";
