+++ assistant/src/schedule/schedule-usage-store.ts
@@ -1,5 +1,5 @@
-import { buildScheduleAttributionSubquery } from "../memory/schedule-attribution-sql.js";
 import { rawAll } from "../persistence/raw-query.js";
+import { buildScheduleAttributionSubquery } from "../persistence/schedule-attribution-sql.js";
 
 export interface ScheduleUsageSummary {
   scheduleId: string;
